+++ R/lock_dir.R
@@ -53,10 +53,29 @@ lock_dir <- function(dir = "./inst/extdata",
     checkmate::assert_flag(remove_file)
     assert_public_key(public_key)
 
+    try_to_lock <- function(file, public_key = "./inst/ssh/id_rsa.pub",
+                         suffix = ".lockr", remove_file = TRUE) {
+        out <- try(
+            lock_file(file, public_key, suffix, remove_file),
+            silent = TRUE
+            )
 
+        if (grepl(paste0(gutils:::escape_regex(suffix), "$"), file)) {
+            "Next"
+        } else if (!inherits(out, "try-error")) {
+            out
+        } else {
+            cli::cli_inform(paste0(
+                "An error occurred while trying to lock ",
+                "'{.strong {cli::col_red(file)}}'."
+            ))
+
+            "Error"
+        }
+    }
 
     vapply(
-        list.files(dir, full.names = TRUE, recursive = TRUE), lock_file,
+        list.files(dir, full.names = TRUE, recursive = TRUE), try_to_lock,
         character(1), public_key = public_key, suffix = suffix,
         remove_file = remove_file
     )
@@ -76,8 +95,32 @@ unlock_dir <- function(dir = "./inst/extdata",
     checkmate::assert_string(password, null.ok = TRUE)
     assert_private_key(private_key, password)
 
+    try_to_unlock <- function(file, private_key = "./inst/ssh/id_rsa",
+                           suffix = ".lockr", remove_file = TRUE,
+                           password = NULL) {
+        out <- try(
+            unlock_file(
+                file, private_key, suffix, remove_file, password
+                ),
+            silent = TRUE
+            )
+
+        if (!grepl(paste0(gutils:::escape_regex(suffix), "$"), file)) {
+            "Next"
+        } else if (!inherits(out, "try-error")) {
+            out
+        } else {
+            cli::cli_inform(paste0(
+                "An error occurred while trying to unlock ",
+                "'{.strong {cli::col_red(file)}}'."
+            ))
+
+            "Error"
+        }
+    }
+
     vapply(
-        list.files(dir, full.names = TRUE, recursive = TRUE), unlock_file,
+        list.files(dir, full.names = TRUE, recursive = TRUE), try_to_unlock,
         character(1), private_key = private_key, suffix = suffix,
         remove_file = remove_file, password = password
         )