+++ NEWS.md
@@ -4,8 +4,7 @@
 
 # encryptrpak (development version)
 
-
-
+* _placeholder_.
 
 # encryptrpak 0.2.0 (2021-11-11)