+++ NEWS.md
@@ -2,6 +2,10 @@
 <!--- https://style.tidyverse.org/news.html -->
 <!--- https://semver.org/ -->
 
+# lockr (development version)
+
+
+
 # lockr 0.3.0 (2023-03-20)
 
 * `encryptrpak` was renamed to `lockr`.